requirements.txt
requests==2.25.1
pytest~=6.2.4
PyYAML==5.4.1
pytest-xdist==2.3.0
deepdiff==5.6.0
pytz==2021.3
pytest-html==3.2.0
pytest-metadata==2.0.4